New modifications

+++ SprintThree/Code/SprintThree/app/src/main/java/one/sprint/alexmalvaez/com/sprintone/database/interfaces/SuperFlingDBInterface.java
@@ -8,7 +8,7 @@ import one.sprint.alexmalvaez.com.sprintone.models.SuperFling;
  * Created by Android1 on 10/20/2015.
  */
 public interface SuperFlingDBInterface {
-    public int addSuperFling(SuperFling superFling);
+    public long addSuperFling(SuperFling superFling);
     public Cursor getSuperFlingById(String id);
     public Cursor getAllSuperFling();
 }
